feat(Switch): add new size
closes #2671
