+++ src/components/Switch/Switch.tsx
@@ -5,48 +5,13 @@ import React, { ChangeEventHandler, useRef } from 'react';
 import { useForkRef } from '../../hooks/useForkRef/useForkRef';
 import { cnMixFocus } from '../../mixs/MixFocus/MixFocus';
 import { cn } from '../../utils/bem';
-import { PropsWithHTMLAttributes } from '../../utils/types/PropsWithHTMLAttributes';
 import { usePropsHandler } from '../EventInterceptor/usePropsHandler';
-
-export const switchPropSize = ['m', 's', 'l'] as const;
-export type SwitchPropSize = typeof switchPropSize[number];
-export const switchPropSizeDefault: SwitchPropSize = switchPropSize[0];
-
-export const switchPropView = ['primary', 'ghost'] as const;
-export type SwitchPropView = typeof switchPropView[number];
-export const switchPropViewDefault: SwitchPropView = switchPropView[0];
-
-export const switchPropAlign = ['center', 'top'] as const;
-export type SwitchPropAlign = typeof switchPropAlign[number];
-export const switchPropAlignDefault: SwitchPropAlign = switchPropAlign[0];
-
-export type SwitchPropOnChange = (object: {
-  e: React.ChangeEvent<HTMLInputElement>;
-  checked: boolean;
-}) => void;
-
-type Props = {
-  checked: boolean | undefined;
-  size?: SwitchPropSize;
-  view?: SwitchPropView;
-  align?: SwitchPropAlign;
-  disabled?: boolean;
-  className?: string;
-  label?: string;
-  onChange?: SwitchPropOnChange;
-  name?: string;
-  onFocus?: React.FocusEventHandler<HTMLInputElement>;
-  onBlur?: React.FocusEventHandler<HTMLInputElement>;
-  autoFocus?: boolean;
-  readOnly?: boolean;
-  required?: boolean;
-  step?: number | string;
-  tabIndex?: number;
-  inputRef?: React.Ref<HTMLInputElement>;
-  children?: never;
-};
-
-export type SwitchProps = PropsWithHTMLAttributes<Props, HTMLLabelElement>;
+import {
+  switchPropAlignDefault,
+  SwitchProps,
+  switchPropSizeDefault,
+  switchPropViewDefault,
+} from './types';
 
 export const cnSwitch = cn('Switch');
