+++ src/components/Switch/__stand__/Switch.dev.stand.mdx
@@ -40,6 +40,7 @@ import { Switch } from '@consta/uikit/Switch';
 За размер переключателя отвечает свойство `size`.
 
 ```tsx
+<Switch size="l" label="Размер XS" checked />
 <Switch size="s" label="Размер S" checked/>
 <Switch size="m" label="Размер М" />
 <Switch size="l" label="Размер L" />
@@ -86,7 +87,7 @@ import { Switch } from '@consta/uikit/Switch';
 | Свойство                  | Тип и значения                                                                  | По умолчанию | Описание                                                                     |
 | ------------------------- | ------------------------------------------------------------------------------- | ------------ | ---------------------------------------------------------------------------- |
 | [`label?`](#название)     | `string`                                                                        | -            | Текст рядом с переключателем                                                 |
-| [`size?`](#размер)        | `'s', 'm', 'l'`                                                                 | `'m'`        | Размер переключателя                                                         |
+| [`size?`](#размер)        | `'xs', 's', 'm', 'l'`                                                           | `'m'`        | Размер переключателя                                                         |
 | [`view?`](#внешний-вид)   | `'primary', 'view'`                                                             | `'primary'`  | Внешний вид                                                                  |
 | [`align?`](#выравнивание) | `'center', 'top'`                                                               | `'center'`   | Выравнивание                                                                 |
 | [`checked`](#состояния)   | `boolean`, `undefined`                                                          | -            | Ставит переключатель в положение **включен**                                 |
