+++ src/components/Switch/index.ts
@@ -1,1 +1,2 @@
 export * from './Switch';
+export * from './types';
